Rename BlockContent/index.jsx
