+++ src/components/BlockContent/blockContent.story.jsx
@@ -1,19 +1,19 @@
 import React from 'react';
-import BlockContent from './index';
+import BlockContent from './BlockContent';
 import Title from '../Title';
 import LeadParagraph from '../LeadParagraph';
 
 export default {
-	title: 'BlockContent',
+    title: 'BlockContent',
 };
 
 export const Default = () => (
-	<div style={{ width: '900px' }}>
-		<BlockContent sectionNumber={1}>
-			<Title noMargin size={1}>
-				Overskrift
-			</Title>
-			<LeadParagraph>Her er det så mye innhold at man bør innlede</LeadParagraph>
-		</BlockContent>
-	</div>
+    <div style={{ width: '900px' }}>
+        <BlockContent sectionNumber={1}>
+            <Title noMargin size={1}>
+                Overskrift
+            </Title>
+            <LeadParagraph>Her er det så mye innhold at man bør innlede</LeadParagraph>
+        </BlockContent>
+    </div>
 );
